requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.1
astroid==2.11.4
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
bandit==1.7.4
billiard==3.6.4.0
black==22.3.0
brotlipy==0.7.0
celery==5.2.6
celery-redbeat==2.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.4
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.13
django-debreach==2.1.0
django-environ==0.8.1
django-extensions==3.1.5
django-import-export==2.8.0
django-js-reverse==0.9.1
django-log-request-id==2.0.0
django-model-utils==4.2.0
django-webpack-loader==1.5.0
djangorestframework==3.13.1
dodgy==0.2.1
dparse==0.5.1
environ==1.0
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.6.0
flake8==3.8.4
flake8-polyfill==1.0.2
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
identify==2.5.0
idna==3.3
ipython==8.3.0
isort==5.10.1
jedi==0.18.1
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupPy==1.14
matplotlib-inline==0.1.3
mccabe==0.6.1
model-bakery==1.5.0
mypy-extensions==0.4.3
nodeenv==1.6.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pbr==5.9.0
pep517==0.12.0
pep8-naming==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==6.6.0
platformdirs==2.5.2
pre-commit==2.19.0
prompt-toolkit==3.0.29
prospector==1.7.7
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.6.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.12.0
pylint==2.13.8
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
requirements-detector==0.7
safety==1.10.3
sentry-sdk==1.5.11
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stack-data==0.2.0
stevedore==3.5.0
tablib==3.2.1
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.1
vulture==2.3
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
